+++ test/runtime/samples/context-api-c/Leaf.svelte
@@ -1,0 +1,7 @@
+<script>
+	import { hasContext } from 'svelte';
+
+	const has = hasContext('test');
+</script>
+
+<div>{has}</div>